requirements.txt
boto==2.49.0
boto3==1.9.74
botocore==1.12.74
certifi==2018.10.15
chardet==3.0.4
Django==1.11.6
docutils==0.14
idna==2.7
jmespath==0.9.3
names==0.3.0
pkg-resources==0.0.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
regex==2018.11.2
requests==2.20.0
s3transfer==0.1.13
six==1.12.0
urllib3==1.24
